+++ cmd/pk.go
@@ -1,0 +1,10 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+var pkCommand = &cobra.Command{
+	Use:   "pk",
+	Short: "Group of package related commands for managing Boilerplate packages.",
+}