+++ cmd/root.go
@@ -1,15 +1,20 @@
 package cmd
 
 import (
+	"errors"
 	"fmt"
-	"os"
-	"path"
-
+	"github.com/charmbracelet/lipgloss"
 	"github.com/oslokommune/ok/cmd/aws"
+	"github.com/oslokommune/ok/cmd/pk"
 	"github.com/oslokommune/ok/cmd/pkg"
+	"github.com/oslokommune/ok/pkg/error_user_msg"
+	"github.com/oslokommune/ok/pkg/pkg/common"
 	"github.com/oslokommune/ok/pkg/pkg/githubreleases"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"os"
+	"path"
+	"strings"
 )
 
 var (
@@ -42,7 +47,21 @@ func Execute() {
 	err := rootCmd.Execute()
 
 	if err != nil {
-		fmt.Println(err)
+		redStyle := lipgloss.NewStyle().
+			Bold(true).
+			Foreground(lipgloss.Color("1")) // Red text
+
+		fmt.Println()
+		fmt.Println(redStyle.Render("Error:"))
+		prettyPrintError(err)
+
+		var userError *error_user_msg.ErrorUserMessage
+		if errors.As(err, &userError) {
+			fmt.Println()
+			fmt.Println(common.StyleTitle.Render("Details:"))
+			fmt.Println(userError.Details())
+		}
+
 		os.Exit(1)
 	}
 }
@@ -57,7 +76,7 @@ func init() {
 
 	// Create dependencies
 	ghReleases := githubreleases.NewGitHubReleases()
-	addCommand := pkg.NewAddCommand()
+	addCommand := pkg.NewAddCommand(ghReleases)
 	updateCommand := pkg.NewUpdateCommand(ghReleases)
 	installCommand := pkg.NewInstallCommand()
 
@@ -74,12 +93,17 @@ func init() {
 	awsCommand.AddCommand(aws.ConfigGeneratorCommand)
 
 	initializeConfiguration()
+
+	if viper.GetBool("enable_experimental") {
+		rootCmd.AddCommand(pkCommand)
+		pkCommand.AddCommand(pk.NewInstallCommand())
+	}
 }
 
 // initializeConfiguration is the function that initializes configuration using viper. It is called at the start of the application.
 func initializeConfiguration() {
 	setConfigFile()
-	viper.SetDefault("enable_experimental", true)
+	viper.SetDefault("enable_experimental", false)
 	viper.SetEnvPrefix("ok")
 	viper.AutomaticEnv()
 	loadConfiguration()
@@ -107,4 +131,44 @@ func setDefaultConfigPath() {
 	viper.AddConfigPath(path.Dir(defaultConfigPath))
 	viper.SetConfigType("yaml")
 	viper.SetConfigName(path.Base(defaultConfigPath))
+}
+
+func prettyPrintError(err error) {
+	i := 0
+
+	for {
+		errStr := err.Error()
+
+		unwrapped := errors.Unwrap(err)
+		if unwrapped == nil {
+			printWithSpaces(errStr, i)
+			break
+		}
+
+		unwrappedStr := unwrapped.Error()
+
+		// As an example, errStr is:
+		// middle b that wraps a: deepest error a
+		//
+		// But we want to print only one error at a time, like this:
+		// middle b that wraps a:
+		//
+		// So we search for "deepest error a" (the unwrapped error) from the complete error string, and remove it.
+		text := strings.Replace(errStr, unwrappedStr, "", 1)
+		printWithSpaces(text, i)
+
+		err = unwrapped
+		i++
+
+		if i > 100 {
+			printWithSpaces("(Too many errors to unwrap, stopping here.)", i)
+			break
+		}
+	}
+}
+
+func printWithSpaces(text string, depth int) {
+	out := strings.Repeat(" ", depth*2) + text
+	fmt.Println(out)
+
 }